@@ -1,4 +1,4 @@
-class AddFullnameToUser < ActiveRecord::Migration
+class AddFullnameToUser < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :fullname, :string
   end
