@@ -1,4 +1,4 @@
-class AddConfirmableToDevise < ActiveRecord::Migration
+class AddConfirmableToDevise < ActiveRecord::Migration[5.2]
   def up
   	add_column :users, :confirmation_token, :string
   	add_column :users, :confirmed_at, :datetime
