@@ -1,4 +1,4 @@
-class AddFieldsToUser < ActiveRecord::Migration
+class AddFieldsToUser < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :provider, :string
     add_column :users, :uid, :string
