@@ -1,4 +1,4 @@
-class AddExtraFieldsToUser < ActiveRecord::Migration
+class AddExtraFieldsToUser < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :phone_number, :string
     add_column :users, :description, :text
