@@ -1,4 +1,4 @@
-class AddAttachmentImageToPhotos < ActiveRecord::Migration
+class AddAttachmentImageToPhotos < ActiveRecord::Migration[5.2]
   def self.up
     change_table :photos do |t|
       t.attachment :image
