@@ -1,4 +1,4 @@
-class AddFieldsToRoom < ActiveRecord::Migration
+class AddFieldsToRoom < ActiveRecord::Migration[5.2]
   def change
     add_column :rooms, :latitude, :float
     add_column :rooms, :longitude, :float
