@@ -12,8 +12,8 @@
 //
 //= require jquery
 //= require bootstrap-sprockets
-//= require jquery_ujs
-//= require jquery-ui/datepicker
+// require jquery_ujs
+// require jquery-ui/datepicker
 //= require toastr
-//= require private_pub
+// require private_pub
 //= require_tree .
