@@ -3,11 +3,13 @@ class ApplicationController < ActionController::Base
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
 
-  before_action :configure_permitted_paramters, if: :devise_controller?
+  before_action :configure_permitted_parameters, if: :devise_controller?
 
   protected
-  	def configure_permitted_paramters
-  		devise_parameter_sanitizer.for(:sign_up) << :fullname
-  		devise_parameter_sanitizer.for(:account_update) << :fullname << :phone_number << :description << :email << :password
-  	end
+  def configure_permitted_parameters
+    # Permit the `subscribe_newsletter` parameter along with the other
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:email, :fullname, :password, :password_confirmation])
+    devise_parameter_sanitizer.permit(:account_update, keys: [:email, :fullname, :phone_number, :description, :password, :password_confirmation, :current_password])
+  end
 end
+
