@@ -42,12 +42,16 @@ Rails.application.configure do
   config.action_mailer.default_url_options = { host: 'localhost', port: 3000 }
 
   config.action_mailer.delivery_method = :smtp
+  # config.action_mailer.smtp_settings = {
+  #   address: 'smtp.mandrillapp.com',
+  #   port: 587,
+  #   enable_starttls_auto: true,
+  #   user_name: 'YOUR_MANDRILL_EMAIL',
+  #   password: 'YOUR_MANDRILL_PASSWORD',
+  #   authentication: 'login'
+  # }
   config.action_mailer.smtp_settings = {
-    address: 'smtp.mandrillapp.com',
-    port: 587,
-    enable_starttls_auto: true,
-    user_name: 'YOUR_MANDRILL_EMAIL',
-    password: 'YOUR_MANDRILL_PASSWORD',
-    authentication: 'login'
+    address: 'localhost',
+    port: 1025
   }
 end
